Move helpers to test_helpers to fix running the token transfer test
SerializeSVMExtraArgs was defined in the messaging test, so it wasn't
available in the token transfer test for some reason
